+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Rename "Reminders" tab to "Weekly" in Wellness ToDo list [#1967](https://github.com/rokwire/illinois-app/issues/1967).
 
 ## [4.1.2] - 2022-07-25
 ### Added